savepoint before altering useStates
